unpack: fast C for unpacking binary flow/packet files.
I noticed that some traces don't have monotonically
increasing timestamps. This can lead to negative interval
values for packets. It's not clear how to handle these,
so I've altered the process program to allow four actions:

  -k : keep them, negative interval values and all (def.)
  -d : discard them entirely, as though they hadn't existed
  -a : take the absolute value of the interval
  -z : use zero for the interval value

The last two leave the packet trace inconsistent in the
sense that the timestamps and intervals recorded will not
match. It's also unclear if the last_time field should be
the last time or the latest time. For now, I'm sticking
with the last.

The unpack program understands and parses the new binary
format for packets, which is less compact because I had
the realization that using 32 bits for intervals wasn't
enough when timestamps have microsecond precision. That
makes the maximum possible interval ~4295 seconds, or
about 72 minutes. We definitely want to be able to have
intervals bigger than that. At the same time, we need to
be able to record negative intervals too. So I simply
gave up on keeping the binary files portable--we just
serialize a double in its native format. Oh, well. Maybe
later I'll come back and fix this.