unpack/pack: added support for autodetection of input type.
This is fairly simple, based on the observation that the IP proto
of a flow is almost never zero, while the first packet in a packet
file will almost always have its most significant byte be zero. In
network byte order, this most significant byte is the very first
byte seen. Thus, we can simply test whether the first character is
the NUL byte to magically determine the type of the input stream.
If the first byte is NUL, then the data is most likely packet data;
if the first byte is non-NUL, then it is most likely flow data.