sortpkts: new syntax for specifying sort orders.
Instead of using different flags for major and minor orders,
now the order of the flags implies whether they are major or
minor: the major field is specified first, followed by minor.

Also, if the minor field is not specified, it is chosen to
jive with the major field. Before it could clash, causing an
error unless both major and minor were given.