unpack: implement index slicing using index list file.
You can slice through the flows or packets with a list of
indices saved in a text file (can be compressed). The output
is identical to what would happen if all the output lines
were saved into an array in memory and then indexed with the
array of indices in the index file.