process & unpack: implemented -D `duplex' flow flag.
This means that flows are treated as being duplex. The
client side is the one with higher source port, or as
a tie-breaker, higher IP address. The server is the
the other side. Outgoing is from client to server, and
incoming is from server to client.

The unpack program will pring outgoing packets as + and
incoming packets as - if the duplex flag is given.

The matrix & pack programs still don't handle duplex.