unpack -R: allow reindexing of flows when using -L.
This works for both flow and packet unpacking. The flows
are implicitly reindexed from zero. If you use the same
index list on a flow file and the corresponding packet
file, you can then repack the outputs to get a new pair
of flow & packet files that match up.

This also works in flow mode, using -T. It does *not*
work in packet mode using -T, since this would require
tracking the order in which we see flows, which is more
than I'm really willing to code up. In either mode, it's
a noop when given the -H flag.