unpack: create -b `binary' output mode.
This mode just outputs the same binary record format that
it takes in. Thus "unpack -b" by itself is a noop. The real
utility of this mode come when you do flow slicing and
renumbering. In those cases, you often don't want to go
through a non-binary form of the data since that's fairly
useless.