histogram: massive change of histogram -o default and semantics.
The offset value is now subtracted from the input value, not
added as it was before. Moreover, the default is now 1, not 0.
This makes histogramming for 1-based integer data work the way
one would expect. An explicit offset of zero must be used for
zero-based data, such as that produced by quantize.