histogram: modulo index mapping and truncation; default is error.
Previously, the default behavior was a mix of modulo mapping for
values that were too large and just ignoring values that were too
small--with the possibility of memory corruption (negative index
into an array).

Now the default is to error out for values not in [0,n-1]. If the
option to use modulo mapping is used, values are mapped using the
mod operation (this is tricky because C's % operator maps negative
numbers to negative numbers). If the option to use truncation is
given, then values that are too small and too large are mapped to
0 and n-1, respectively.