bin/pairflows: perl script to pair up flows into full-duplexes.
The flow appearing first in the flows file is listed first; since
this flow had the first packet appearing in the original trace,
it is justifiable to consider the first flow as the client-to-
server flow of the two, and the second flow as the server-to-
client flow. Flows with no pair are listed at the end on a line
with no other flow id.